requirements.txt
pandas
numpy
openai==0.28.0
faiss-cpu
